requirements.txt
paho-mqtt
posix_ipc
pyudev
spidev
pyserial
